pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Generate diff"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "PyYAML>=6.0.1"
]
requires-python = ">=3.12"

[project.scripts]
gendiff = "gendiff.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9"
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
